requirements.txt
tensorflow
sklearn
pillow
scipy
